Moved user entity to entity folder and updated column types
