+++ src/auth/get-user.decorator.ts
@@ -1,6 +1,6 @@
 import { createParamDecorator } from '@nestjs/common';
 
-import { User } from './user.entity';
+import { User } from './entity/user.entity';
 
 export const GetUser = createParamDecorator((data, { args: [argsElement] }): User => {
   return argsElement.user;
